pyproject.toml
[project]
name = "hedwig"
version = "2.0.0"
description = "Self-Evolving Personal AI Signal Radar — algorithm sovereignty for individuals"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.0",
    "httpx>=0.27",
    "supabase>=2.0",
    "slack-sdk>=3.0",
    "feedparser>=6.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
]

[project.scripts]
hedwig = "hedwig.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
